+++ src/fleet/fleetActions/bombard/BombardHandler.cpp
@@ -29,9 +29,10 @@ namespace bombard
 				this->one = rand() % 101;
 				this->two = config.nget("ship_bomb_factor",1) + (config.nget("ship_bomb_factor",0) * this->tLevel + ceil(this->shipCnt / 10000) + this->f->getSpecialShipBonusBuildDestroy() * 100);
 				
-				if (this->one < this->two) {
+				if (this->one < this->two) 
+				{
 					// level the building down, at least one level 
-					this->bLevel = ceil(this->shipCnt/2500.0);
+					this->bLevel = (short)ceil(this->shipCnt/2500.0);
 					
 					std::string actionString = this->targetEntity->bombBuilding(this->bLevel);
